+++ routes/api.js
@@ -58,6 +58,7 @@ const transferAdmin = require("./projects/transferAdmin");
 const script = require("./projects/script");
 
 const updateLabels = require("./labelling/updateLabels");
+const deleteLabels = require("./labelling/deleteLabels");
 
 const downloadDataset = require("./downloads/downloadDataset");
 const downloadProject = require("./downloads/downloadProject");
@@ -109,6 +110,7 @@ api.post("/script", script);
 
 // LABELLING ROUTES
 api.post("/updateLabels", updateLabels);
+api.delete("/deleteBadLabels/:Admin/:PName/:Lid", deleteLabels);
 
 // DOWNLOAD ROUTES
 api.post("/downloadDataset", downloadDataset);
@@ -879,278 +881,6 @@ api.put("/api/switchLabels", async (req, res) => {
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
-});
-
-api.delete("/deleteBadLabels/:Admin/:PName/:Lid", async (req, res) => {
-    try {
-        const Admin = req.params.Admin;
-        const PName = req.params.PName;
-        const Lid = req.params.Lid.split(",");
-
-        console.log("Admin: ", Admin);
-        console.log("PName: ", PName);
-        console.log("Lid: ", Lid);
-
-        const public_path = __dirname.replace("routes", ""),
-            main_path = public_path + "public/projects/",
-            project_path = main_path + Admin + "-" + PName;
-
-        const dbPath = project_path + "/" + PName + ".db";
-
-        console.log("this is the dbPath: ", dbPath);
-
-        const db = new sqlite3.Database(dbPath, (err) => {
-            if (err) {
-                console.error(err.message);
-                return res.status(500).send("Database connection error");
-            }
-            console.log(`Connected to ${PName} database`);
-        });
-
-        console.log("Lid:", Lid);
-        const placeholders = Lid.map(() => "?").join(",");
-        console.log("placeholders", placeholders);
-        const sql = `DELETE FROM Labels WHERE LID IN (${placeholders})`;
-
-        db.run(sql, Lid, function (err) {
-            if (err) {
-                console.error(err.message);
-                return res.status(500).send("Internal server error");
-            }
-            if (this.changes === 0) {
-                return res.status(404).send("Product not found");
-            }
-
-            console.log(`Deleted entry with ID: ${Lid}`);
-            return res.status(200).json({
-                message: `This transaction was deleted for ID: ${Lid}`,
-            });
-        });
-
-        db.close((err) => {
-            if (err) {
-                console.error(err.message);
-            } else {
-                console.log("Database connection closed");
-            }
-        });
-    } catch (error) {
-        console.error(error);
-        if (!res.headersSent) {
-            res.status(500).json({ message: error.message });
-        }
-    }
-});
-
-api.post("/batch-change-class", async (req, res) => {
-    console.log("Batch Change Class");
-    var project_name = req.body.PName,
-        admin = req.body.Admin,
-        class1 = req.body.class1,
-        class2 = req.body.class2;
-
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/", // $LABELING_TOOL_PATH/public/projects/
-        project_path = main_path + admin + "-" + project_name; // $LABELING_TOOL_PATH/public/projects/project_name
-
-    var aidb = new sqlite3.Database(
-        project_path + "/" + project_name + ".db",
-        (err) => {
-            if (err) {
-                return console.error(err.message);
-            }
-            console.log("Connected to aidb.");
-        },
-    );
-    aidb.getAsync = function (sql) {
-        var that = this;
-        return new Promise(function (resolve, reject) {
-            that.get(sql, function (err, row) {
-                if (err) {
-                    console.log("runAsync ERROR! ", err);
-                    //reject(err);
-                    aidb.close(function (err) {
-                        if (err) {
-                            console.error(err);
-                        } else {
-                            console.log("aidb closed");
-                        }
-                    });
-                    reject(err);
-                } else resolve(row);
-            });
-        }).catch((err) => {
-            console.error(err);
-            // res.send({"Success": "No"})
-            return res.send("ERROR! " + err);
-        });
-    };
-    aidb.allAsync = function (sql) {
-        var that = this;
-        return new Promise(function (resolve, reject) {
-            that.all(sql, function (err, row) {
-                if (err) {
-                    console.log("runAsync ERROR! ", err);
-                    //reject(err);
-                    aidb.close(function (err) {
-                        if (err) {
-                            console.error(err);
-                        } else {
-                            console.log("aidb closed");
-                        }
-                    });
-                    reject(err);
-                } else resolve(row);
-            });
-        }).catch((err) => {
-            console.error(err);
-            // res.send({"Success": "No"})
-            return res.send("ERROR! " + err);
-        });
-    };
-    aidb.runAsync = function (sql) {
-        var that = this;
-        return new Promise(function (resolve, reject) {
-            that.run(sql, function (err, row) {
-                if (err) {
-                    console.log("runAsync ERROR! ", err);
-                    aidb.close(function (err) {
-                        if (err) {
-                            console.error(err);
-                        } else {
-                            console.log("aidb closed");
-                        }
-                    });
-                    reject(err);
-                } else resolve(row);
-            });
-        }).catch((err) => {
-            console.error(err);
-            // res.send({"Success": "No"})
-            return res.send("ERROR! " + err);
-        });
-    };
-
-    await aidb.runAsync(
-        "UPDATE Labels SET CName = '" +
-            class2 +
-            "' WHERE CName ='" +
-            class1 +
-            "'",
-    );
-    await aidb.runAsync(
-        "UPDATE Validation SET CName = '" +
-            class2 +
-            "' WHERE CName ='" +
-            class1 +
-            "'",
-    );
-    res.send({ Success: "Yes" });
-});
-
-api.post("/solo-change-class", async (req, res) => {
-    var LID = parseInt(req.body.LID),
-        selectedClass = req.body.selectedClass,
-        project_name = req.body.PName,
-        admin = req.body.Admin;
-
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/", // $LABELING_TOOL_PATH/public/projects/
-        project_path = main_path + admin + "-" + project_name; // $LABELING_TOOL_PATH/public/projects/project_name
-
-    var aidb = new sqlite3.Database(
-        project_path + "/" + project_name + ".db",
-        (err) => {
-            if (err) {
-                return console.error(err.message);
-            }
-            console.log("Connected to aidb.");
-        },
-    );
-    aidb.getAsync = function (sql) {
-        var that = this;
-        return new Promise(function (resolve, reject) {
-            that.get(sql, function (err, row) {
-                if (err) {
-                    console.log("runAsync ERROR! ", err);
-                    //reject(err);
-                    aidb.close(function (err) {
-                        if (err) {
-                            console.error(err);
-                        } else {
-                            console.log("aidb closed");
-                        }
-                    });
-                    reject(err);
-                } else resolve(row);
-            });
-        }).catch((err) => {
-            console.error(err);
-            // res.send({"Success": "No"})
-            return res.send("ERROR! " + err);
-        });
-    };
-    aidb.allAsync = function (sql) {
-        var that = this;
-        return new Promise(function (resolve, reject) {
-            that.all(sql, function (err, row) {
-                if (err) {
-                    console.log("runAsync ERROR! ", err);
-                    //reject(err);
-                    aidb.close(function (err) {
-                        if (err) {
-                            console.error(err);
-                        } else {
-                            console.log("aidb closed");
-                        }
-                    });
-                    reject(err);
-                } else resolve(row);
-            });
-        }).catch((err) => {
-            console.error(err);
-            // res.send({"Success": "No"})
-            return res.send("ERROR! " + err);
-        });
-    };
-    aidb.runAsync = function (sql) {
-        var that = this;
-        return new Promise(function (resolve, reject) {
-            that.run(sql, function (err, row) {
-                if (err) {
-                    console.log("runAsync ERROR! ", err);
-                    aidb.close(function (err) {
-                        if (err) {
-                            console.error(err);
-                        } else {
-                            console.log("aidb closed");
-                        }
-                    });
-                    reject(err);
-                } else resolve(row);
-            });
-        }).catch((err) => {
-            console.error(err);
-            // res.send({"Success": "No"})
-            return res.send("ERROR! " + err);
-        });
-    };
-
-    await aidb.runAsync(
-        "UPDATE Labels SET CName = '" +
-            selectedClass +
-            "' WHERE LID =' " +
-            LID +
-            "'",
-    );
-    await aidb.runAsync(
-        "UPDATE Validation SET CName = '" +
-            selectedClass +
-            "' WHERE LID =' " +
-            LID +
-            "'",
-    );
-    res.send({ Success: "Yes" });
 });
 
 module.exports = api;
